requirements.txt
imutils==0.5.1
numpy==1.15.1
opencv-python==3.4.3.18
Pillow==5.2.0
PyAutoGUI==0.9.38
PyMsgBox==1.0.6
PyScreeze==0.1.18
pytesseract==0.2.4
PyTweening==1.0.3
